+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Features
+
+- Add ability to forward incoming heartbeats to another Wakatime compatible server.
+
 ## [1.3.2] - 2021-04-11
 
 ### Bug fixes